Addong DTO's for preventing accidental exposure of unnecessary data.

+++ src/main/java/com/enviro/assessment/grad001/KaraboMashao/model/DisposalGuideline.java
@@ -24,7 +24,7 @@ public class DisposalGuideline {
 
 
     // Getters and Setters
-    public int getDisposalGuidelineId() {
+    public long getDisposalGuidelineId() {
         return disposalGuidelineId;
     }
